arm: Build Depthcharge FIT partition images for supported boards
On most ARM chromebooks the Chromium OS bootloader ("Depthcharge") wants
its OS kernel, initramfs and device-tree to be packaged as a FIT image
inside its own format. We can package U-Boot in this way to make
Depthcharge chainload into it by a process currently explained per board
in doc/chromium/chainload.rst. This patch generalizes and automates
building this image using binman.

Depthcharge on 32-bit ARM boards doesn't copy the kernel out of the FIT
before booting it. If U-Boot is not built as position-independent it
must be placed carefully in the FIT image so that Depthcharge puts it
exactly where it's built to run from.

So far, this placement was done by changing SYS_TEXT_BASE to known-good
values which happen to be 0x100 offset from this memory region, and
having board-specific *.its files which add some whitespace in the image
description to ensure their correct position in the FIT.

This patch generalizes this by inserting an empty padding image into the
FIT whose size depends on the SYS_TEXT_BASE value, so that changing it
will still place U-Boot in the correct place.

Enable building images for these boards, by setting necessary config
options and including the binman definition in their device-trees.

The KERNEL_START and KERNEL_SIZE values are from Depthcharge's sources,
but from different git branches of it that correspond to vendor releases
of the board firmwares:

  Board             | Branch (firmware-*) | KERNEL_START | KERNEL_SIZE
  ------------------+---------------------+--------------+-------------
  chromebook_speedy | veyron-6588.B       | 0x02000000   | 0x01000000
  chromebook_minnie |  "                  |  "           |  "
  chromebook_jerry  |  "                  |  "           |  "
  chromebit_mickey  |  "                  |  "           |  "
  chromebook_bob    | gru-8785.B          | 0x15000000   | 0x02000000
  chromebook_kevin  |  "                  |  "           |  "
  nyan-big          | nyan-5771.B         | 0x81000000   | 0x01000000

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>