arm: Build Depthcharge partition images for supported boards
On ARM chromebooks the Chromium OS bootloader ("Depthcharge") wants its
OS kernel and device-tree to be packaged as a FIT image inside its own
format. We can package U-Boot in this way to make Depthcharge chainload
into it by a process currently explained in doc/chromium/chainload.rst.

Automate building this kind of image with binman for applicable boards,
by setting necessary config options, adding a common binman definition
of this image and adding it in the boards' device-trees.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>