rockchip: kevin: Enable configs for the display
The kevin board uses an Embedded DisplayPort panel compatible with the
simple panel and Rockchip eDP drivers. Its backlight is controlled by
the Chromium OS Embedded Controller Pulse Width Modulator. Enable these
for the board.

Also set VIDEO_ROCKCHIP_MAX_{XRES,YRES} to 2400x1600, the resolution of
its panel. The defaults of 1920x1080 would be too low but setting them
to 3840x2160 doesn't seem to work.

The Rockchip video driver needs to assert/deassert some resets, so also
enable the reset controller. RESET_ROCKCHIP defaults to y for this board
when DM_RESET=y, so it's enough to set that.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>