rockchip: kevin: Allow setting up clocks in U-Boot proper
The vendor firmware for the kevin board sets the clocks differently from
what U-Boot uses. When chainloading U-Boot proper from the vendor
firmware we need to reinitialize the clocks for (at least) the display
to work properly.

This was already implemented by using the SPL handoff mechanism to check
if U-Boot (not the vendor firmware) has initialized the clocks. Enable
the configs necessary to trigger this reinitialization.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>