test/py: Use sudo mount if guestmount fails in filesystem tests
If guestmount isn't available on the system, filesystem test setup falls
back to using sudo mount to prepare its disk images. If guestmount is
available but fails to work, the tests are immediately skipped. Instead
of giving up on a guestmount failure, try using sudo mount as an attempt
to keep tests running.

Also don't check if guestmount is in PATH, as trying to run a missing
guestmount will follow the same failure codepath and fall back to sudo
mount anyway.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>