test/py: Move filesystem test helper functions to utils module
The filesystem test setup has functions that would be useful to some EFI
tests (capsule and secure boot), as both create pre-populated filesystem
images to be used in their tests. Move these to the utility functions
module as a first step of unifying filesystem image creation.

The "tool_is_in_path" function in the filesystem tests was only used to
check if guestmount is available. Since the check is removed it has no
other users there and can be removed. It also exists in the global
conftest.py and tests might want to use it in the future, so unify it
there as well instead of just removing it.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>