rockchip: gru: Set up SoC IO domain registers
The RK3399 SoC needs to know the voltage value provided by some
regulators, which is done by setting relevant register bits. Configure
these the way other RK3399 boards do, but with values set in coreboot.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>
Commit-notes:
There is a driver for this on Rockchip's repo, I managed to forward-port
it and get it working. If that's more desirable than doing it per-board
like this I can send that upstream (but I'd prefer to do it after this).
END