video: Add cursor support for DM_VIDEO consoles
So far the DM_VIDEO console is completely lacking any cursor, which makes
typing and correcting quite irritating.

Add a simple cursor display by writing a SPACE glyph in the background
colour to the next character position on the screen. Any typed character
will naturally overwrite it, so we need to only explicitly clear it if
the next character will appear somewhere else (newline, backspace).

Link: https://lore.kernel.org/r/20220110005638.21599-2-andre.przywara@arm.com
Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
[Alper: Rebase for console_set_font()]
Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>