video: Kconfig: convert CONFIG_VIDEO_FONT_4X6 to Kconfig
We used to have two boards using a very tiny font, replacing the
standard 8x16 font used on most framebuffers. This was done outside of
Kconfig though, so move this over now.

As those boards have been removed lately, there are currently no users,
but we will gain more font support in a later patch.

Fix the build for the 4x6 font on the way.

Link: https://lore.kernel.org/u-boot/20220110005638.21599-4-andre.przywara@arm.com/
Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Reviewed-by: Simon Glass <sjg@chromium.org>