binman: Make FIT etype a subclass of Section etype
Binman tries to expand some entries into parts that make it up, e.g.
'u-boot' into a 'u-boot-expanded' section that contains 'u-boot-nodtb'
and 'u-boot-dtb'. Entries with child entries must call ExpandEntries()
on them to build a correct image, as it's possible that unexpanded child
entries have no data of their own. The FIT entry type doesn't currently
do this, which means putting u-boot inside it doesn't work.

It's rather simple to implement ExpandEntries() for FIT, but it makes
more sense to turn it into a subclass of Section which already has it,
do that instead.

Add a copy of a simple FIT image test to check subentry expansion in FIT
entries, which used to be broken before this. Also fix the subentry
alignment test which now attempts to write symbols to a nonexistent SPL
ELF test file.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>