Kconfig: Add options related to chainloading from Depthcharge
The Chromium OS bootloader ("Depthcharge") can chainload U-Boot in a few
ways. Although the details vary from board to board, in general these
are: as a "kernel" partition image, as a "Legacy Boot" payload, and as
an "Alternative Firmware" payload. Future patches will automate building
these images using binman, but U-Boot must also satisfy some constraints
to work in these configurations.

On ARM64 boards chainloading from a partition, Depthcharge checks
for the Linux kernel image header magic and copies the binary-to-run to
a dynamically chosen address, so we must build U-Boot as position
independent and with the Linux header.

On 32-bit ARM boards chainloading from a partition, U-Boot must either
be built position-independent or built to run from somewhere in the
board-specific fixed memory region Depthcharge loads it to.

For chainloading from a "Legacy Boot" or "Alternative Firmware" image on
ARM boards, we need to build an ELF file including a device-tree binary
to put in a CBFS image.

Add configuration options we can use to enable/disable building these
images and enforce these constraints.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>