arm: Build Depthcharge uImage partition images for supported boards
The very first ARM chromebooks had a modified U-Boot as their firmware
which implements the Chromium OS verified boot flows and can have its OS
kernel and device-tree packed as a uImage inside their own format. We
can package U-Boot in this way to chainload into a newer build that is
not restricted to this verified boot mechanism. This patch automates
building such an image using binman.

Enable building images for these boards, by setting the necessary config
options and including the binman definition in their device-trees.

The KERNEL_START and KERNEL_SIZE values are derived from Chromium OS'
U-Boot sources, but from different git branches of it that correspond to
vendor releases of the board firmwares:

  Board             | Branch (firmware-*) | KERNEL_START | KERNEL_SIZE
  ------------------+---------------------+--------------+-------------
  snow              | snow-2695.B         | 0x42000000   | 0x00800000
  spring            | spring-3824.B       | 0x42000000   | 0x00800000
  peach-pit         | pit-4482.B          | 0x20008000   | 0x00800000
  peach-pi          | pit-4482.B          | 0x20008000   | 0x00800000

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>