HACK: rockchip: bob: Enable REMAKE_ELF
This builds a u-boot.elf file from u-boot.bin which has an appended dtb.
We can use this ELF as the "Legacy Boot" CBFS payload for Depthcharge.

Binman support that builds this image is incomplete. For now, enable the
prerequisite config as a HACK so I can manually build the CBFS later as:

    cbfstool u-boot.elf.cbfs create -s "$((2 * 1024 * 1024)" -m arm64
    cbfstool u-boot.elf.cbfs add-payload -f u-boot.elf -n payload -c lzma

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>