binman: Change how faked blobs are created
At present fake blobs are created but internally an empty blob is used.
Change it to use the contents of the faked file. Also return whether the
blob was faked, in case the caller needs to know that.

Add a TODO to put fake blobs in their own directory.

Series-changes: 2
- Add a patch to change how faked blobs are created

Signed-off-by: Simon Glass <sjg@chromium.org>